[FIX] Encoding tuples of static content tuples
Container type is only 32 bytes size (head) when it's dynamic. Arrays
are always dynamic, but tuples are only encoded in dymanic if any of the
contained values are dynamic.

In the case where there's a tuple of tuples with static content, then
container size is encoded in place, and the size (head) is the static size of
the contained elements together.